Updated for admin query

+++ src/main/java/io/sankha/entity/Skill.java
@@ -2,11 +2,6 @@ package io.sankha.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -26,49 +21,49 @@ public class Skill {
 
   @Column(name = "angular", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long angular;
+  private int angular;
 
   @Column(name = "react", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long react;
+  private int react;
 
   @Column(name = "spring", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long spring;
+  private int spring;
 
   @Column(name = "restful", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long restful;
+  private int restful;
 
   @Column(name = "hibernate", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long hibernate;
+  private int hibernate;
 
   @Column(name = "git", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long git;
+  private int git;
 
   @Column(name = "docker", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long docker;
+  private int docker;
 
   @Column(name = "jenkins", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long jenkins;
+  private int jenkins;
 
   @Column(name = "aws", nullable = false, table = "skill")
   @Size(max = 20, min = 0, message = "criteria not met")
-  private Long aws;
+  private int aws;
 
   @Column(name = "spoken", nullable = false, table = "skill")
-  @Size(max=20,min=0,message="criteria not met")
-  private Long spoken;
+  @Size(max = 20, min = 0, message = "criteria not met")
+  private int spoken;
 
   @Column(name = "communication", nullable = false, table = "skill")
-  @Size(max=20,min=0,message="criteria not met")
-  private Long communication;
+  @Size(max = 20, min = 0, message = "criteria not met")
+  private int communication;
 
   @Column(name = "aptitude", nullable = false, table = "skill")
-  @Size(max=20,min=0,message="criteria not met")
-  private Long aptitude;
+  @Size(max = 20, min = 0, message = "criteria not met")
+  private int aptitude;
 }
